requirements.txt
fastapi
uvicorn
langchain
langgraph
groq
huggingface-hub
pydantic
sqlalchemy
psycopg2-binary
python-dotenv
pytest
langchain_groq
langchain_core
redis
fakeredis
